nxp-lx2xxx: update u-boot version and drop optimizing bootargs
Issue: LIN10-4958

Update the u-boot version and drop the special optimizing bootargs setting which has
documented on the DPDK chapter. Testing found an information about RCU, which is
overloaded during the test that RCU cannot get handled in time. After removing the
related default parameters, the RCU information is not found anymore. So drop the
optimizing parameters by default.

Signed-off-by: Xulin Sun <xulin.sun@windriver.com>